Responsive

+++ src/app/post-list/post-list.component.ts
@@ -5,6 +5,7 @@ import { map } from 'rxjs/operators';
 import { PostServiceService } from '../post-service.service';
 import { Post } from '../models/post.model';
 import { MatAccordion } from '@angular/material/expansion';
+import Swal from 'sweetalert2';
 
 @Component({
   selector: 'app-post-list',
@@ -22,10 +23,10 @@ export class PostListComponent implements OnInit {
   constructor(private postService: PostServiceService, private router: Router) {}
 
   ngOnInit() {
-    this.categories.forEach(category => {
+    this.categories.forEach((category) => {
       this.selectedCategories[category] = false;
     });
-    this.applyFilters();
+    this.applyFilters(); // Initial filtering
   }
 
   applyFilters() {
@@ -37,20 +38,47 @@ export class PostListComponent implements OnInit {
             post.content.toLowerCase().includes(this.searchText.toLowerCase())) &&
           (Object.keys(this.selectedCategories).some(
             (category) => this.selectedCategories[category] && post.category === category
-          ) || Object.keys(this.selectedCategories).every(
-            (category) => !this.selectedCategories[category]
-          ))
+          ) ||
+            Object.keys(this.selectedCategories).every(
+              (category) => !this.selectedCategories[category]
+            ))
         )
       )
     );
   }
-  
+
+  onSearchTextChange() {
+    this.applyFilters();
+  }
+
+  onCategoryChange() {
+    this.applyFilters();
+  }
 
   editPost(post: Post) {
     this.router.navigate(['/postform'], { queryParams: { id: post.id } });
   }
 
   deletePost(postId: number) {
-    this.postService.deletePost(postId);
+    Swal.fire({
+      title: 'Are you sure?',
+      text: 'Do you really want to delete this post? This action cannot be undone!',
+      icon: 'warning',
+      showCancelButton: true,
+      confirmButtonColor: '#3085d6',
+      cancelButtonColor: '#d33',
+      confirmButtonText: 'Yes, delete it!',
+    }).then((result) => {
+      if (result.isConfirmed) {
+        this.postService.deletePost(postId);
+        Swal.fire({
+          title: 'Deleted!',
+          text: 'The post has been deleted.',
+          icon: 'success',
+          confirmButtonText: 'OK',
+        });
+        this.applyFilters(); // Refresh filtered posts after deletion
+      }
+    });
   }
 }
